feat: implement autocomplete (#4)
Implements an autocomplete in the header.

The autocomplete features recent searches and Query Suggestions. When submitting the form, it navigates to the search page with a `query` URL search parameter. When the `query` search parameter is present, it's used as the initial value of the input.

Since we're using the `render` React API to render the autocomplete, we lose access to all the React contexts, including the router one from Next.js. For this reason, I wrapped the `AutocompleteItem` component with `RouterContext.Provider` (an undocumented import from Next.js) so we can pass down the router to it and use `<Link>` components inside.